# src/main/java/com/gsuitesafe/backup/dto/CreatedBackupResponse.java
package com.gsuitesafe.backup.dto;

import java.util.UUID;

public class CreatedBackupResponse {

    private final UUID backupId;

    public CreatedBackupResponse(UUID backupId) {
        this.backupId = backupId;
    }

    public UUID getBackupId() {
        return backupId;
    }
}