# src/main/resources/static/api-docs-swagger.yaml
openapi: 3.0.2

info:
  title: Backup API
  description: API used to retrieve mocked backups
  version: 1.0.0
paths:
  /backups:
    post:
      tags:
        - Backup
      summary: Creates a backup
      description: Creates a backup
      operationId: createBackup
      responses:
        200:
          description: Backup created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedBackupResponse"
              example:
                backupId: 3160c9de-b152-4886-ae52-41f670c493e9
        405:
          description: Method not allowed error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodNotAllowedErrorResponse"
    get:
      tags:
        - Backup
      summary: Retrieves list of initiated backups
      description: Retrieves a list of the initiated backups
      operationId: listInitiatedBackups
      responses:
        200:
          description: Initiated backup list retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListOfInitiatedBackups"
        405:
          description: Method not allowed error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodNotAllowedErrorResponse"
  /exports/{backupId}:
    get:
      tags:
        - Export
      summary: Generates a zip file containing backup
      description: Generates a zip file containing backup
      operationId: downloadFullBackup
      parameters:
        - name: backupId
          in: path
          description: Backup id of the backup that will we zipped
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Backup retrieved successfully
          content:
            application/zip:
              schema:
                type: string
                format: binary

        405:
          description: Method not allowed error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodNotAllowedErrorResponse"
  /exports/{backupId}/{label}:
    get:
      tags:
        - Export
      summary: Generates a zip file containing backup by label
      description: Generates a zip file containing backup
      operationId: downloadBackupByLabel
      parameters:
        - name: backupId
          in: path
          description: Backup id of the backup that will we zipped
          required: true
          schema:
            type: integer
        - name: label
          in: path
          description: Label of the messages in the backup that will we zipped
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Backup retrieved successfully
          content:
            application/zip:
              schema:
                type: string
                format: binary

        405:
          description: Method not allowed error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodNotAllowedErrorResponse"
components:
  schemas:
    CreatedBackupResponse:
      type: object
      properties:
        backupId:
          type: string
          description: Backup id
          example: "3160c9de-b152-4886-ae52-41f670c493e9"
    ListOfInitiatedBackups:
      type: array
      items:
        type: object
        properties:
          backupId:
            type: string
            description: Backup id
          date:
            type: string
            description: Creation backup date
          status:
            type: string
            description: Backup status
      example:
        - backupId: "3160c9de-b152-4886-ae52-41f670c493e9"
          date: "2021-07-17"
          status: "OK"
        - backupId: "ff3f1fb9-0645-4b42-97ed-0429c2d707d2"
          date: "2021-07-17"
          status: "In progress"
    MethodNotAllowedErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          description: timestamp of the response
          example: "2020-05-30T21:09:11.897+0000"
        status:
          type: integer
          description: http status code
          example: 405
        error:
          type: string
          description: Description of the error
          example: "Method Not Allowed"
        message:
          type: string
          description: Message of the response
          example: ""
        path:
          type: string
          description: Path called
          example: "/endpoint"